Lock down react-overlays version
any version of react-overlays above 0.7.3 breaks our build when we attempt to remove modals during our tests. Normally, it would not be a problem to override the dependency in our npm-shrinkwrap, but unmet peer dependencies of redux-form and react-md-spinner are preventing us from effectively using npm-shrinkwrap. Also, partial npm-shrinkwraps are not working for a reason that I cannot determine. When that is eventually straightened out, we can go back to using the mainstream version of react-bootstrap.